Avoid a warning with Sun CC

+++ platform.mk
@@ -6,6 +6,15 @@ GCC_CLANG=$(shell $(CC) -v 2>&1 | \
   grep -q -i -e "gcc version" -e "clang version" 2> /dev/null && \
   printf '%s\n' "1")
 
+# Detect if CC is Sun CC with `-V`
+SUNCC_CMP=$(shell $(CC) -V 2>&1 | \
+  grep -q -e "Sun C" 2> /dev/null && printf '%s\n' "1")
+
+# Sun CC: Disable LTO
+ifeq ($(SUNCC_CMP),1)
+ NO_LTO:=1
+endif
+
 # Detect if CC is GCC by name
 ifneq "$(findstring gcc,$(CC))" ""
  GCC_CLANG=1
@@ -23,31 +32,33 @@ ifeq ($(GCC_CLANG),1)
 endif
 
 # Detect if CC supports `-flto`
-FLTO_WR:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
-  $(CC) -flto .test.c -o .test.out > /dev/null 2>&1; \
-    echo $$?; rm -f .test.c .test.out > /dev/null 2>&1)
-ifeq ($(FLTO_WR),0)
- FLTO_OK:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
-   $(CC) -Werror -flto .test.c -o .test.out > /dev/null 2>&1; \
+ifndef NO_LTO
+ FLTO_WR:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
+   $(CC) -flto .test.c -o .test.out > /dev/null 2>&1; \
      echo $$?; rm -f .test.c .test.out > /dev/null 2>&1)
- ifeq ($(FLTO_OK),0)
-  LTO_FLAGS:=-flto
-  # Detect if CC supports `-flto=auto`
-  AUTO_WR:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
-    $(CC) -flto=auto .test.c -o .test.out > /dev/null 2>&1; \
+ ifeq ($(FLTO_WR),0)
+  FLTO_OK:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
+    $(CC) -Werror -flto .test.c -o .test.out > /dev/null 2>&1; \
       echo $$?; rm -f .test.c .test.out > /dev/null 2>&1)
-  ifeq ($(AUTO_WR),0)
-   AUTO_OK:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
-     $(CC) -Werror -flto=auto .test.c -o .test.out > /dev/null 2>&1; \
+  ifeq ($(FLTO_OK),0)
+   LTO_FLAGS:=-flto
+   # Detect if CC supports `-flto=auto`
+   AUTO_WR:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
+     $(CC) -flto=auto .test.c -o .test.out > /dev/null 2>&1; \
        echo $$?; rm -f .test.c .test.out > /dev/null 2>&1)
-   ifeq ($(AUTO_OK),0)
-    LTO_FLAGS:=-flto=auto
+   ifeq ($(AUTO_WR),0)
+    AUTO_OK:=$(shell printf '%s\n' "int main(void){return 0;}" > .test.c; \
+      $(CC) -Werror -flto=auto .test.c -o .test.out > /dev/null 2>&1; \
+        echo $$?; rm -f .test.c .test.out > /dev/null 2>&1)
+    ifeq ($(AUTO_OK),0)
+     LTO_FLAGS:=-flto=auto
+    endif
    endif
   endif
  endif
+ CFLAGS+=$(LTO_FLAGS)
+ LDFLAGS+=$(LTO_FLAGS)
 endif
-CFLAGS+=$(LTO_FLAGS)
-LDFLAGS+=$(LTO_FLAGS)
 
 # Solaris or illumos: Force `-flto=auto` to `-flto`
 ifneq "$(findstring SunOS,$(UNAME_S))" ""
